feat(UI Design): Moved style objects into a separate file
