+++ src/components/AbstractUserField/style.tsx
@@ -1,0 +1,17 @@
+import type * as CSS from 'csstype';
+
+export const CotainerStyle: CSS.Properties = {
+  display: 'flex',
+  flexDirection: 'column',
+  gap: '4px',
+};
+
+export const LabelStyle: CSS.Properties = {
+  display: 'block',
+};
+
+export const InputStyle: CSS.Properties = {
+  height: '24px',
+  padding: '4px 8px',
+  fontSize: '16px',
+};
