+++ src/components/Button/index.tsx
@@ -1,21 +1,10 @@
-import type * as CSS from 'csstype';
+import { ButtonStyle } from './style';
 
 interface ButtonProps {
   id?: string;
   name: string;
   handleClick: () => void;
 }
-
-const ButtonStyle: CSS.Properties = {
-  color: '#666',
-  fontSize: '16px',
-  backgroundColor: '#eee',
-  border: '2px solid #ddd',
-  borderRadius: '8px',
-  padding: '4px 8px',
-  cursor: 'pointer',
-  minWidth: '52px',
-};
 
 export function Button({ id, name, handleClick }: ButtonProps) {
   return (
