+++ src/components/Button/style.tsx
@@ -1,0 +1,12 @@
+import type * as CSS from 'csstype';
+
+export const ButtonStyle: CSS.Properties = {
+  color: '#666',
+  fontSize: '16px',
+  backgroundColor: '#eee',
+  border: '2px solid #ddd',
+  borderRadius: '8px',
+  padding: '4px 8px',
+  cursor: 'pointer',
+  minWidth: '52px',
+};
