+++ src/components/Layout/style.tsx
@@ -1,0 +1,14 @@
+import type * as CSS from 'csstype';
+
+export const ContainerStyle: CSS.Properties = {
+  display: 'flex',
+  flexDirection: 'column',
+  gap: '32px',
+  minWidth: '400px',
+  maxWidth: '600px',
+  margin: '0px auto',
+};
+
+export const TitleStyle: CSS.Properties = {
+  margin: '0px 0px 8px 0px',
+};
