+++ src/components/LoginForm/style.tsx
@@ -1,0 +1,15 @@
+import type * as CSS from 'csstype';
+
+export const LoginFormStyle: CSS.Properties = {
+  display: 'flex',
+  flexDirection: 'column',
+  maxWidth: '300px',
+  gap: '16px',
+  padding: '32px 16px',
+  backgroundColor: 'darkgray',
+};
+
+export const LoginButtonStyle: CSS.Properties = {
+  height: '48px',
+  marginTop: '4px',
+};
