+++ src/components/TodoCreateForm/index.tsx
@@ -1,25 +1,10 @@
-import type * as CSS from 'csstype';
-
 import { useState } from 'react';
-import { Button } from './Button';
+import { Button } from '../Button';
+import { ContainerStyle, ModifyInputStyle } from './style';
 
 export interface CreateFormProps {
   onAddNewTodo: (name: string) => void;
 }
-
-const ContainerStyle: CSS.Properties = {
-  display: 'flex',
-  flexDirection: 'row',
-  gap: '16px',
-  width: '100%',
-};
-
-const ModifyInputStyle: CSS.Properties = {
-  fontSize: '18px',
-  lineHeight: '1.5',
-  padding: '4px 8px',
-  width: '100%',
-};
 
 export function TodoCreateForm({ onAddNewTodo: addNewTodo }: CreateFormProps) {
   const [createInput, setCreateInput] = useState('');
