+++ src/components/TodoCreateForm/style.tsx
@@ -1,0 +1,15 @@
+import type * as CSS from 'csstype';
+
+export const ContainerStyle: CSS.Properties = {
+  display: 'flex',
+  flexDirection: 'row',
+  gap: '16px',
+  width: '100%',
+};
+
+export const ModifyInputStyle: CSS.Properties = {
+  fontSize: '18px',
+  lineHeight: '1.5',
+  padding: '4px 8px',
+  width: '100%',
+};
