+++ src/components/TodoListItem/index.tsx
@@ -1,7 +1,12 @@
-import type * as CSS from 'csstype';
-
 import { useState } from 'react';
-import { Button } from './Button';
+import { Button } from '../Button';
+import {
+  CheckboxStyle,
+  ContainerStyle,
+  LabelStyle,
+  ModifyInputStyle,
+  NameStyle,
+} from './style';
 
 interface TodoListItemProps {
   name: string;
@@ -23,34 +28,6 @@ export function TodoListItem({
   const commitModified = async () => {
     await handleSubmit(input);
     setModifyMode(false);
-  };
-
-  const ContainerStyle: CSS.Properties = {
-    display: 'flex',
-    gap: '16px',
-    alignItems: 'center',
-  };
-
-  const LabelStyle: CSS.Properties = {
-    width: '240px',
-    display: 'flex',
-    gap: '8px',
-  };
-
-  const CheckboxStyle: CSS.Properties = {
-    minWidth: '18px',
-    accentColor: '#333',
-  };
-
-  const ModifyInputStyle: CSS.Properties = {
-    fontSize: '18px',
-    lineHeight: '1.5',
-    padding: '4px 8px',
-    width: '100px',
-  };
-
-  const NameStyle: CSS.Properties = {
-    fontStyle: 'italic',
   };
 
   return (
