+++ src/components/TodoListItem/style.tsx
@@ -1,0 +1,29 @@
+import type * as CSS from 'csstype';
+
+export const ContainerStyle: CSS.Properties = {
+  display: 'flex',
+  gap: '16px',
+  alignItems: 'center',
+};
+
+export const LabelStyle: CSS.Properties = {
+  width: '240px',
+  display: 'flex',
+  gap: '8px',
+};
+
+export const CheckboxStyle: CSS.Properties = {
+  minWidth: '18px',
+  accentColor: '#333',
+};
+
+export const ModifyInputStyle: CSS.Properties = {
+  fontSize: '18px',
+  lineHeight: '1.5',
+  padding: '4px 8px',
+  width: '100px',
+};
+
+export const NameStyle: CSS.Properties = {
+  fontStyle: 'italic',
+};
